pyproject.toml
[project]
name = "pyfunckit"
version = "0.0.0"
description = "PYthon FUNCtion ToolKIT: Common code for knative faas"
requires-python = ">=3.13"
dependencies = [
    "pydantic==2.11.3",
    "pyjwt==2.10.1",
    "instructor==1.7.9",
    "flask",
]
